migrate agenda input to emotion to get rid of warning
